Use t.Cleanup() and t.parallel()
