+++ database/helper_test.go
@@ -2,18 +2,49 @@ package database
 
 import (
 	"database/sql"
+	"errors"
+	"os"
+	"testing"
+
+	"github.com/golang-migrate/migrate/v4"
+
+	_ "github.com/golang-migrate/migrate/v4/database/mysql"
+	_ "github.com/golang-migrate/migrate/v4/source/file"
+	_ "github.com/golang-migrate/migrate/v4/source/github"
+
+	"gorm.io/gorm"
 
 	"github.com/masibw/blog-server/config"
 	"github.com/masibw/blog-server/log"
 	"gorm.io/driver/mysql"
-	"gorm.io/gorm"
 	"moul.io/zapgorm2"
 )
+
+var db *gorm.DB
+
+func TestMain(m *testing.M) {
+	var err error
+	var mig *migrate.Migrate
+	mig, err = migrate.New("file://"+os.Getenv("MIGRATION_FILE"), "mysql://"+config.PureDSN())
+	if err != nil {
+		panic(err)
+	}
+	if err = mig.Up(); err != nil {
+		if !errors.Is(err, migrate.ErrNoChange) {
+			panic(err)
+		}
+	}
+
+	db = NewTestDB()
+	code := m.Run()
+	os.Exit(code)
+}
 
 func NewTestDB() *gorm.DB {
 	logger := zapgorm2.New(log.GetPureLogger())
 	logger.SetAsDefault()
-	db, err := gorm.Open(mysql.Open(config.DSN()), &gorm.Config{Logger: logger})
+	var err error
+	db, err = gorm.Open(mysql.Open(config.DSN()), &gorm.Config{Logger: logger})
 	if err != nil {
 		panic(err.Error())
 	}
